imed_v1/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.5
django-filter==23.5
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docopt==0.6.2
et_xmlfile==2.0.0
idna==3.10
mod_wsgi==5.0.2
numpy==2.3.2
openpyxl==3.1.5
pandas==2.3.1
pipreqs==0.4.13
psycopg==3.2.9
psycopg-binary==3.2.9
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
wheel==0.45.1
yarg==0.1.10
